requirements.txt
browser-history
mega.py